# [Performance] optimise cloneVNode routine