--- a/benchmarks/uibench/index.html
+++ b/benchmarks/uibench/index.html
@@ -8,7 +8,7 @@
 <body>
 <div id="App"></div>
 <script src="https://localvoid.github.io/uibench-base/0.1.0/uibench.js"></script>
-<script src="dist/inferno.min.js"></script>
+<script src="dist/inferno.js"></script>
 <script src="app.js" type="text/javascript"></script>
 </body>
 </html>