This PR improves cloneVNode performance. 

Performance is achieved by using createVNode instead of Object.assign to clone vnode.
Object assign also has to copy "**dom**" property of old node while createVNode keeps it null.

Performance comparison (npm run pretest:bench):
BEFORE:
Single node nulls at **21372733 ops/sec**
Single node not defined at **22574141 ops/sec**
Regular shape at **152152 ops/sec**
20 children siblings at **40055 ops/sec**
33 nested nodes at **646924 ops/sec**
Table shape at **36119 ops/sec**
navigation shape ul - li at **176940 ops/sec**
typical html page at **59197 ops/sec**
deeply nested Arrays (nobody should do this) at **246104 ops/sec**

AFTER:
Single node nulls at **20717821 ops/sec**
Single node not defined at **22540122 ops/sec**
Regular shape at **709036 ops/sec**
20 children siblings at **227014 ops/sec**
33 nested nodes at **637949 ops/sec**
Table shape at **162090 ops/sec**
navigation shape ul - li at **715051 ops/sec**
typical html page at **238392 ops/sec**
deeply nested Arrays (nobody should do this) at **573518 ops/sec**

current implementation of dbmonster also shows mean render time -1ms