@lukeed 2017-01-01
This is awesome 😄 I only wonder why both `Single Node` tests experienced some slow down. Can someone enlighten me?