@Havunen 2017-01-01
@lukeed the numbers are so large so even minor background stuff affects it (error margin).